UPSTREAM: soc/intel/skylake: Refactor PEG configuration
Simplify some if-blocks which are used for the configuration, enablement
and disablement of the PEG devices.

This changes the logic of the code, since it configures PegxEnable
before the if-blocks, where x is the number of the PEG device, and the
further configuration of the PEG devices depends on the enablement of
PegxEnable.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ib335ad694836fc2ae7b9dbdf8fc069efa72cb46d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e32fa4e152b212bf1d886b2007b01521a7161cbc
Original-Change-Id: I6dd88ce752ce8f0255c424d0e5b2d8ef918885a1
Original-Signed-off-by: Felix Singer <felixsinger@posteo.net>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44368
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Michael Niewöhner
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2360624
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>