UPSTREAM: soc/intel/skylake: Factor out unnecessary if-else-block
Move InternalGfx config option out of the if-else-block and replace the
left over config option IgdDvmt50PreAlloc by a ternary expression. Also,
adjust related code comments to fit the new logic of this code.

This changes the logic of the code, since InternalGfx is configured
first and IgdDvmt50PreAlloc depends on its value. The negation in the
ternary expression is removed to improve the readability.

BUG=none
BRANCH=none
TEST=none

Change-Id: I2a169d6f6326a6ea6c845a6d5b2f8209218fc557
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 736de9f2466a08d563cc9ed29c76d9d5f0cc2440
Original-Change-Id: I89ff17f4574a7ade228c1791f17ea072fb731775
Original-Signed-off-by: Felix Singer <felixsinger@posteo.net>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44369
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Michael Niewöhner
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2360623
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>