UPSTREAM: mb/google/zork: Disable ACP I2S wake for schematic version 3.6+
Starting with v3.6 of reference schematics, headphone jack interrupt
is moved to a standard GPIO instead of using CODEC_GPI. Thus, we no
longer need I2S wake to be enabled in the ACP for boards using v3.6+
version of schematics.

This change sets `acp_i2s_wake_enable` and `acp_pme_enable` to default
0 in baseboard devicetrees and overrides to 1 in update_hp_int_odl()
if the board is still using older version of reference schematics.

BUG=b:159934887

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: fd8840880d5eeac2db19d3295d33155c6e2557e0
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Change-Id: I44b40db95b5148fe483c7340c5bd0d58627970a7
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44403
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: If5ed0f30fad9c3430094416acc6cb6d0d627c47e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2354976
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>