UPSTREAM: mb/google/asurada: fix EC commands timeout
The Asurada EC is using the large packet (256B) mode, and we were
seeing lots of timeout errors on various commands.

The AcceptTimeoutUs in EC SPI driver is hard-coded at 5000,
and that is too small for large packet running in 1M so we
should change EC SPI to the same value that kernel is using (3M).

BUG=b:161509047
TEST=emerge-asurada coreboot chromeos-bootimage; flash and boot

Change-Id: I0b34dae86a710ab08ff68bccc45aa3eabee8e263
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6871d51125dcf7c091c9c461c3e9b8bfa681217d
Original-Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Original-Change-Id: I9c47324022129ca23ef75d0c80e215da1692636d
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/46394
Original-Reviewed-by: Yu-Ping Wu <yupingso@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2498397
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>