UPSTREAM: soc/amd/picasso: Reboot for recovery if no psp workbuf is found
Instead of halting if the vboot workbuf is not passed to coreboot by the
PSP, reset and reboot into recovery mode.

This process is made more difficult because if the workbuf isn't
available, we can't reboot directly into recovery - the workbuf is
needed for that process to be done through the regular calls, and we
don't want to go around the vboot API and just write into VBNV directly.
To overcome this, we set a CMOS flag, and reset the system.
PSP_verstage checks for this flag so it will update VBNV and reset the
system after generating the workbuf.

BUG=b:152638343
TEST=Simulate the workbuf not being present and verify the reboot
process.

Change-Id: Ia9130953958262e26c82cf77fa0c783121fcb5d0
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 50cca76e546bc5528be28ff99e717982d2783356
Original-Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Change-Id: I049db956a5209904b274747be28ff226ce542316
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44538
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2375140
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>