UPSTREAM: util: Add support to spd_tools for fixed id
For boards that have already assigned memory ids, there needs to be a
way to fix parts to a specific id. After assigning all the fixed ids the
tool still attempts to minimize the SPDs entries. Since a fixed ID could
be anywhere, gaps can be created in the list. So an empty SPD entry is
created to fill the gaps in the list until they are used.

BUG=b:162939176
TEST=Generate various outputs

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: a662648a7fbfc04bb0f4a2ef92d7faa4fb8c7e09
Original-Signed-off-by: Rob Barnes <robbarnes@google.com>
Original-Change-Id: I1f8ea1ff4f33a97ab28ba94896a1054e89189576
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44463
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Change-Id: I51797c6d38bf0964963cc6262398683ce9c3d02e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2382140
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>