UPSTREAM: mb/ti/beaglebone: Load romstage/ramstage from SD
Adds a "sd_media" boot_device to allow booting from the SD card. This
assumes that the generated "MLO" file is placed at a 128KB offset from
the start of the SD card, to allow for the MBR etc. to be at the start
of the SD card. Placing the MLO file here allows the AM335x boot ROM to
load and execute the bootblock stage as well, as 128KB is one of the
offsets the boot ROM checks when looking for the next stage to execute.

As part of this, a FMD for the Beaglebone has also been defined. It's
sized at 32M somewhat arbitrarily, as SD cards could allow for much
bigger payloads.

TEST: Beaglebone boots from bootblock into romstage. Romstage to
ramstage still doesn't work as it needs RAM initialization first.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: c25d54b97e0f6debf37e25841f53114b6318063d
Original-Change-Id: I5f6901217fb974808e84aeb679af2f47eeae30fd
Original-Signed-off-by: Sam Lewis <sam.vr.lewis@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44385
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: I35ca34561aac64965c36af82924e26b3ac22ae5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2795721
Reviewed-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>