UPSTREAM: superio/ite: delay PWRGD3 during resume
According to the IT8728F datasheet it is possible to add an extra delay
between 3VSBSW# being set and PWRGD3 being set during resume from
Suspend-to-RAM. This is enabled in the special function selection
register, the default being 0.

This is also useful for the IT8720F although this chip does not have the
PWRGD3 output. On the corresponding pin it has PWROK2, which the setting
then seems to apply to.

The datasheet for the IT8720F marks the corresponding bit as reserved,
but the vendor BIOS of an Acer Aspire M3800 sets it anyway. Without
setting the bit, coreboot fails to resume from S3. Oscilloscope
measurements have shown that setting the bit increases the delay between
3VSBSW# being set and PWROK2 being set from around 1 us to 140 ms. The
actual use of PWROK2 on the board design is unclear - the only
destination it seems to reach is a pin header near the SuperIO marked as
"GPIO1".

BUG=none
BRANCH=none
TEST=none

Change-Id: I7bcd7fe7690046dc1d78b9042dbb760b4ad9a882
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6c5f47b50ce8a65ccf3039e109af3740572c5345
Original-Signed-off-by: Michael Büchler <michael.buechler@posteo.net>
Original-Change-Id: I51cbf2470dc2b840a647a20090acb5a0cf4f4025
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44639
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2372242
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>