UPSTREAM: lib/imd: Fix imdr_recover for small regions
One of the checks inside imdr_recover() was written with the
assumption that imdr limit is always aligned to LIMIT_ALIGN. This is
true only for large allocations, thus may fail for small regions.

It's not necessary to check if root_pointer is under the limit, since
this is implicitly verified by imdr_get_root_pointer().

BUG=none
BRANCH=none
TEST=none

Change-Id: Ibeb5a067cedadd65dde147f669d077ee24164b9b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 086c5e6fc866e99208da3c1c9c8840cd0aeaec80
Original-Signed-off-by: Jan Dabros <jsd@semihalf.com>
Original-Change-Id: I25d6291301797d10c6a267b5f6e56ac38b995b7b
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44667
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2377966
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>