UPSTREAM: libpayload: memmove: Don't make expectations of architecture memcpy
default_memmove() calls memcpy() when (src > dst). This is safe for the
default_memcpy() implementation, but just calling memcpy() may invoke an
architecture-specific implementation. Architectures are free to
implement memcpy() however they want and may assume that buffers don't
overlap in either direction. So while this happens to work for all
current architecture implementations of memcpy(), it's safer not to rely
on that and only rely on the known implementation of default_memcpy()
for the forwards-overlapping case.

BUG=none
BRANCH=none
TEST=none

Change-Id: I7d8340d9c8e8e2d81a9608959ac6f4ad37b12cd6
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ae096be00c3ada5acc6dfd601a1ad2bb36e234db
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I7ece4ce9e6622a36612bfade3deb62f351877789
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44691
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2372278
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>