UPSTREAM: soc/amd/picasso: copy local info to transfer buf
We added transfer_info_struct to contain various information about
memory region we pass from PSP to x86 in commit 0c12abe462.

This should be at the start of transfer region but we only manipulated
it as local variable and didn't put data into the region, resulting
garbage data for transfer_info when x86 tries to read it.

Copy the content of local variable to beginning of _transfer_buffer
before requesting transfer to PSP so coreboot on x86 can access it.

BUG=b:159220781
BRANCH=zork
TEST=check transfer_info_struct is correctly populated on romstage

Change-Id: Id2051a8437e52c190f953129bfad861d04ed2bfc
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 5f027fa4c227c9788a4b7d88d3bdf3891ed6d844
Original-Signed-off-by: Kangheui Won <khwon@chromium.org>
Original-Change-Id: I14bc34e6af501240a6f633db3999a7759e88d60b
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44751
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2409801
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>