UPSTREAM: cpu/x86/smm/smmhandler: Fix x86_64 assembly exit
Fix an issue the assembler didn't warn about to fix a crash on real
hardware. qemu didn't catch this issue either.

The linker uses the same address for variables in BSS if they aren't
initialized in the code. This results in %edx being set to the value
of %eax, which causes an exception restoring IA32_EFER on real
hardware.

Tested on qemu with KVM enabled.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic5ddec130e7ad7dac53dde54138ddf439e4870c2
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 9256e51f48550c83f5425550b8feee9fc2b7366c
Original-Change-Id: Ie36a88a2a11a6d755f06eff9b119e5b9398c6dec
Original-Signed-off-by: Patrick Rudolph <siro@das-labor.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44780
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2383661
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>