UPSTREAM: symbols: Change implementation details of DECLARE_OPTIONAL_REGION()
It seems that GCC's LTO doesn't like the way we implement
DECLARE_OPTIONAL_REGION(). This patch changes it so that rather than
having a normal DECLARE_REGION() in <symbols.h> and then an extra
DECLARE_OPTIONAL_REGION() in the C file using it, you just say
DECLARE_OPTIONAL_REGION() directly in <symbols.h> (in place and instead
of the usual DECLARE_REGION()). This basically looks the same way in the
resulting object file but somehow LTO seems to like it better.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: 644a512e560147324ecf74ebce5e336bc57dd7bf
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I6096207b311d70c8e9956cd9406bec45be04a4a2
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44791
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Original-Reviewed-by: HAOUAS Elyes <ehaouas@noos.fr>
Change-Id: I3a5e5795e4b6153dbf1dd5518005cf5888d324f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2382136
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>