UPSTREAM: xhci: Fix Abort command event handling
This fixes issues with how we handle events generated by the xHCI
"command abort" command. first, depending on the state of the xHCI
controller, the COMMAND_ABORTED may not be generated. If the
controller was between commands, only the COMMAND_RING_STOPPED event
will be generated. Second, do not adjust the command ring "cur"
pointer as that just confuses the controller.

BUG=b:160354585,b:157123390
TEST=able to boot into recovery using USB stick on servo v2 on volteer
        as well as HooToo 8-1 hub

Change-Id: Ibb89bdca2185fb3915a31db768d526593f65615d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 551216a4d1a52a56a9de4859873694cbcb0e6109
Original-Change-Id: I055df680d1797f35d9730e2bfdb4119925657168
Original-Signed-off-by: Caveh Jalali <caveh@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44875
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2384888
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>