UPSTREAM: util/gen_spd: translate DeviceBusWidth to die bus width
If a memory part is a x16 part that has two dies and only a single
rank, then the x16 describes the part width (since this solution will
need to be a stacked solution) and as such, we must translate the
DeviceBusWidth to the "die bus width" instead.

Change DeviceBusWidth variable name to PackageBusWidth to be more
descriptive

BUG=b:166645306, b:160157545
TEST=run gen_spd and verify that spds for parts matching description
above changed appropriately.

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: 913ea9278f50b73427a50a54e91f8d5502d93219
Original-Change-Id: Ia6f3ca109d344b7a015da28125a94ce10d2bdfb8
Original-Signed-off-by: Nick Vaccaro <nvaccaro@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44870
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Change-Id: I10adbe9d98de95d96ffb94413d444a972500951f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2382145
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>