UPSTREAM: soc/amd/picasso/southbridge.h: rename GPP clock setting offsets
The _SHIFT postfix is a bit clearer than the _SHL one and more in line
with the names used for this kind of defines in coreboot. The
documentation on that register is currently wrong and will hopefully be
fixed in the future; the defines should now match the hardware.

BUG=b:149970243
BRANCH=zork

Change-Id: I7e3e7daaa2c83450e58b9f8b64f88ff3dda1ddfa
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 05ef94795f0303326084290690659f99c6c0e9a7
Original-Change-Id: I977f107d466521484ca13fa1f4dd86a50c8150d7
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44888
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2384897
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>