UPSTREAM: memory_info: add max_speed_mts and configured_speed_mts
ddr_frequency is ambiguous and is interpreted differently in several
places. Instead of renaming this field, this deprecates it and adds
two new fields with unambiguous naming, max_speed_mts and
configured_speed_mts. smbios.c falls back to using ddr_frequency
when either of these fields are 0.

The same value was being used for both configured memory speed and
max memory speed in SMBIOS type 17, which is not accurate when
configured speed is not the max speed.

BUG=b:167218112
TEST=Boot ezkinil, no change to dmidecode -t17

Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: 327f1058d2feaa022407128aa9dac55f408b714c
Original-Change-Id: Iaa75401f9fc33642dbdce6c69bd9b20f96d1cc25
Original-Signed-off-by: Rob Barnes <robbarnes@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/44549
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: I5c31d717d35d09b87597bf4eca7cf295be397b8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2393112
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>