UPSTREAM: common/block/pmc: Add a check to program the PchPmPwrCycDur
This patch adds a check to avoid violating the PCH EDS recommendation
that the PchPmPwrCycDur will never be smaller than the the SLP_Sx
assertion widths.

This code was initially added for cannonlake and now moving it to common
code since the same check will be used to program the PchPmPwrCycDur
for Jasperlake and Tigerlake.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia6b26edf5242c2cd91bc4323a93063478af39d5a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 186250f68e432f0c32b0c9909be6008cbc6aca47
Original-Change-Id: Ie7d5f54939c5eb1f885d303f75a04958b9d77f4d
Original-Signed-off-by: V Sowmya <v.sowmya@intel.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/45028
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2411978
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>