UPSTREAM: mb/google/volteer: Change default camera power GPIO to 0
The default GPIO values for camera power were set as 1 so the LED was
turned on by default when the board is powered on.
This status is kept until the camera is probed then being turned off.
So the LED is turned on for a few seconds during the boot up.

By setting the default power to 0, the LED is lit only when camera
is turned on for probing and this should be just a blink.

BUG=b:167635396
BRANCH=none
TEST=Build and boot volteer board. Monitor camera privacy LED
and check it is not lit more than 0.5 seconds.

Change-Id: I4022c3de360461e576ac68ebd1a3e4a57c93ba7b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 029069960e4568d0100f67fffe815f1ff035df80
Original-Signed-off-by: Daniel Kang <daniel.h.kang@intel.com>
Original-Change-Id: Ic7df391aa512daafe6e1ce49e9222b90e17ad806
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/45058
Original-Reviewed-by: Caveh Jalali <caveh@chromium.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2440811
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>