UPSTREAM: mb/google/zork: Set eMMC presets
They should be tuned per board to get the best signal and boot time.

This fixes the HS400 preset, so it's correctly set to A. It also changes
the SDR50 and DDR50 presets to B. We can't boot correctly when DDR50 is
set to A.

I chose 1 as the init kHz value since that's what depthcharge uses to
calculate the init clock.

BUG=b:159823235
TEST=Boot Ezkinil and dump SDHCI preset registers.

Change-Id: If69d2518cb53807870ecceb0fc8971d93361faea
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 94be1f7399629de0578c56625ddd327cc1123fe4
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Change-Id: Ie2f3497b65d771820ab1a803fec73265547f8906
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/45098
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2436140
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>