UPSTREAM: program_loading: Replace prog_rdev() with raw start pointer and size
Since prog_locate() was eliminated, prog_rdev() only ever represents the
loaded program in memory now. Using the rdev API for this is unnecessary
if we know that the "device" is always just memory. This patch changes
it to be represented by a simple pointer and size. Since some code still
really wants this to be an rdev, introduce a prog_chain_rdev() helper to
translate back to that if necessary.

BUG=none
BRANCH=none
TEST=none

Change-Id: I3c11ac5282246cb0a3b6d95b3d6b3cdaca02b4b4
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2e973942bc34ff2d7b110ba35bf3cda987838907
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: If7c0f1c5698fa0c326e23c553ea0fe928b25d202
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/46483
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2766750
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>